@@ -6,27 +6,29 @@ import { CommonModule } from '@angular/common';
 import { BrowserAnimationsModule } from '@angular/platform-browser/animations';
 import { StoreModule } from '@ngrx/store';
 import { EffectsModule } from '@ngrx/effects';
-import { AuthEffects } from './store/effects/auth.effects';
+import { AuthEffects } from '@src/app/store/effects/auth.effects';
 import { StoreRouterConnectingModule, RouterStateSerializer } from '@ngrx/router-store';
-import { reducer, metaReducers } from './store/reducers';
-import { CustomRouterStateSerializer } from './shared/utils';
-import { AppRoutes } from './app-routing.module';
-import { AppComponent } from './app.component';
-import { CustMaterialModule } from './shared/cust-material/cust-material.module';
-import { EllipsisPipe } from './pipes/ellipsis.pipe';
-import { BookService } from './services/book.service';
-import { SearchBooksComponent } from './pages/search-books/search-books.component';
-import { BookPreviewComponent } from './components/book-preview/book-preview.component';
-import { BooksListComponent } from './components/books-list/books-list.component';
+import { reducer, metaReducers } from '@src/app/store/reducers';
+import { CustomRouterStateSerializer } from '@src/app/shared/utils';
+import { AppRoutes } from '@src/app/app-routing.module';
+import { AppComponent } from '@src/app/app.component';
+import { CustMaterialModule } from '@src/app/shared/cust-material/cust-material.module';
+import { EllipsisPipe } from '@src/app/pipes/ellipsis.pipe';
+import { BookService } from '@src/app/services/book.service';
+import { SearchBooksComponent } from '@src/app/pages/search-books/search-books.component';
+import { BookPreviewComponent } from '@src/app/components/book-preview/book-preview.component';
+import { BooksListComponent } from '@src/app/components/books-list/books-list.component';
 import { FormsModule, ReactiveFormsModule } from '@angular/forms';
-import { HttpClientModule } from '@angular/common/http';
-import { BookDetailPageComponent } from './pages/book-detail-page/book-detail-page.component';
-import { BookDetailsComponent } from './components/book-details/book-details.component';
+import { HttpClientModule, HTTP_INTERCEPTORS } from '@angular/common/http';
+import { BookDetailPageComponent } from '@src/app/pages/book-detail-page/book-detail-page.component';
+import { BookDetailsComponent } from '@src/app/components/book-details/book-details.component';
 import { LocationStrategy, HashLocationStrategy } from '@angular/common';
-import { LoginModal } from './components/Auth/login-modal.component';
+import { LoginModal } from '@src/app/components/Auth/login-modal.component';
 import { CdkTableModule } from '@angular/cdk/table';
-import { AuthService } from './services/auth.service';
+import { AuthService } from '@src/app/services/auth.service';
 import {catchError,exhaustMap,map,filter,switchMap,retry,finalize} from 'rxjs/operators';
+import { TokenInterceptor } from '@src/app/core/token.interceptor';
+import { AutoGeneratedComponent } from '@src/app/auto-generated/auto-generated.component';
 
 @NgModule({
   declarations: [
@@ -37,13 +39,14 @@ import {catchError,exhaustMap,map,filter,switchMap,retry,finalize} from 'rxjs/op
     BooksListComponent,
     BookDetailPageComponent,
       BookDetailsComponent,
-      LoginModal
+      LoginModal,
+      AutoGeneratedComponent
   ],
     imports: [
         BrowserAnimationsModule,
         BrowserModule, CommonModule,
         StoreModule.forRoot(reducer,{ metaReducers }),
-        StoreRouterConnectingModule, CdkTableModule,
+        StoreRouterConnectingModule.forRoot(), CdkTableModule,
     RouterModule,
     AppRoutes,
         FormsModule, ReactiveFormsModule,
@@ -51,13 +54,14 @@ import {catchError,exhaustMap,map,filter,switchMap,retry,finalize} from 'rxjs/op
     // BooksModule,
         CustMaterialModule,
         EffectsModule.forRoot([AuthEffects]),
-        StoreRouterConnectingModule.forRoot()
+       
   ],
   providers: [{
         provide: LocationStrategy,
         useClass: HashLocationStrategy
   },
       { provide: RouterStateSerializer, useClass: CustomRouterStateSerializer },
+      {provide:HTTP_INTERCEPTORS,useClass:TokenInterceptor,multi:true},
       BookService,
       AuthService
     ],
