@@ -1,5 +1,5 @@
 import { Component } from '@angular/core';
-import { MatDialog } from '@angular/material';
+import { MatDialog } from '@angular/material/dialog';
 import { LoginModal } from './components/Auth/login-modal.component';
 import { Observable } from 'rxjs';
 @Component({
